// opencdx-dashboard/src/api/communications/model/message-template.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Timestamp } from './timestamp';

/**
 * 
 * @export
 * @interface MessageTemplate
 */
export interface MessageTemplate {
    /**
     * 
     * @type {string}
     * @memberof MessageTemplate
     */
    'templateId'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof MessageTemplate
     */
    'variables'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof MessageTemplate
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageTemplate
     */
    'content'?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageTemplate
     */
    'type'?: Type;
    /**
     * 
     * @type {Timestamp}
     * @memberof MessageTemplate
     */
    'created'?: Timestamp;
    /**
     * 
     * @type {Timestamp}
     * @memberof MessageTemplate
     */
    'modified'?: Timestamp;
    /**
     * 
     * @type {string}
     * @memberof MessageTemplate
     */
    'creator'?: string;
    /**
     * 
     * @type {string}
     * @memberof MessageTemplate
     */
    'modifier'?: string;
}

export const Type = {
    UnspecifiedMessageType: 'UNSPECIFIED_MESSAGE_TYPE',
    Info: 'INFO',
    Warning: 'WARNING',
    Error: 'ERROR',
    Unrecognized: 'UNRECOGNIZED'
} as const;

export type Type = typeof Type[keyof typeof Type];